@@ -1,251 +1,141 @@
 {
-    "name": "MapStore-C179",
-    "version": "1.0.0",
-    "description": "PUBLIACQUA",
-    "repository": "git@github.com:geosolutions-it/MapStore-C179.git",
-    "main": "index.js",
-    "devDependencies": {
-        "@babel/core": "7.5.5",
-        "@babel/plugin-proposal-class-properties": "7.5.5",
-        "@babel/plugin-syntax-dynamic-import": "7.2.0",
-        "@babel/preset-env": "7.5.5",
-        "@babel/preset-react": "7.0.0",
-        "@geosolutions/acorn-jsx": "4.0.2",
-        "@geosolutions/jsdoc": "3.4.4",
-        "@geosolutions/mocha": "6.2.1-3",
-        "axios-mock-adapter": "1.16.0",
-        "babel-eslint": "10.0.3",
-        "babel-istanbul-loader": "0.1.0",
-        "babel-loader": "8.0.5",
-        "babel-plugin-add-module-exports": "0.1.4",
-        "babel-plugin-object-assign": "1.2.1",
-        "babel-plugin-react-transform": "2.0.2",
-        "cesium": "1.17.0",
-        "copy-webpack-plugin": "5.0.2",
-        "css-loader": "0.19.0",
-        "denodeify": "1.2.1",
-        "docma": "1.5.1",
-        "download-cli": "1.0.1",
-        "dynamic-public-path-webpack-plugin": "1.0.4",
-        "escope": "3.2.0",
-        "eslint": "6.3.0",
-        "eslint-plugin-no-only-tests": "2.3.1",
-        "eslint-plugin-react": "3.3.2",
-        "expect": "1.20.1",
-        "file-loader": "2.0.0",
-        "glob": "7.1.1",
-        "html-loader": "0.5.1",
-        "html-webpack-plugin": "3.2.0",
-        "jsdoc-jsx": "0.1.0",
-        "karma": "4.3.0",
-        "karma-chrome-launcher": "3.1.0",
-        "karma-cli": "2.0.0",
-        "karma-coverage": "2.0.1",
-        "karma-coveralls": "2.1.0",
-        "karma-firefox-launcher": "1.2.0",
-        "karma-ie-launcher": "1.0.0",
-        "karma-junit-reporter": "1.2.0",
-        "karma-mocha": "1.3.0",
-        "karma-mocha-reporter": "2.2.5",
-        "karma-sourcemap-loader": "0.3.7",
-        "karma-webpack": "4.0.2",
-        "less": "2.7.1",
-        "less-loader": "4.1.0",
-        "mini-css-extract-plugin": "0.5.0",
-        "mkdirp": "0.5.1",
-        "ncp": "2.0.0",
-        "parallelshell": "1.2.0",
-        "postcss": "7.0.14",
-        "postcss-loader": "3.0.0",
-        "postcss-prefix-selector": "1.7.1",
-        "progress-bar-webpack-plugin": "1.12.1",
-        "raw-loader": "0.5.1",
-        "react-motion": "0.5.0",
-        "react-transition-group": "1.1.3",
-        "readline-promise": "1.0.4",
-        "redux-devtools": "3.4.0",
-        "redux-devtools-dock-monitor": "1.1.2",
-        "redux-devtools-log-monitor": "1.3.0",
-        "redux-immutable-state-invariant": "2.1.0",
-        "redux-mock-store": "1.2.2",
-        "rimraf": "2.5.2",
-        "simple-git": "1.33.1",
-        "style-loader": "0.12.4",
-        "url-loader": "0.5.7",
-        "vusion-webfonts-generator": "0.4.1",
-        "webpack": "4.41.5",
-        "webpack-bundle-size-analyzer": "2.0.2",
-        "webpack-cli": "3.3.6",
-        "webpack-dev-server": "3.7.2"
-    },
-    "//": "replace react-sortable-items with official on npm when it will support React 15",
-    "dependencies": {
-        "@boundlessgeo/jsonix": "2.4.3",
-        "@carnesen/redux-add-action-listener-enhancer": "0.0.1",
-        "@geosolutions/proj4": "2.4.7",
-        "@geosolutions/react-joyride": "1.10.2",
-        "@geosolutions/wkt-parser": "1.2.2",
-        "@mapbox/geojsonhint": "2.0.1",
-        "@mapbox/togeojson": "0.16.0",
-        "@turf/bbox": "4.1.0",
-        "@turf/bbox-polygon": "5.1.5",
-        "@turf/boolean-contains": "5.1.5",
-        "@turf/boolean-overlap": "5.1.5",
-        "@turf/center": "5.1.5",
-        "@turf/circle": "6.0.1",
-        "@turf/great-circle": "5.1.5",
-        "@turf/inside": "4.1.0",
-        "@turf/line-intersect": "4.1.0",
-        "@turf/point-on-surface": "4.1.0",
-        "@turf/polygon-to-linestring": "4.1.0",
-        "ag-grid-community": "20.2.0",
-        "ag-grid-react": "20.2.0",
-        "axios": "0.18.1",
-        "b64-to-blob": "1.2.19",
-        "babel-polyfill": "6.8.0",
-        "babel-standalone": "6.7.7",
-        "bootstrap": "3.3.5",
-        "canvas-to-blob": "0.0.0",
-        "canvg-browser": "1.0.0",
-        "chroma-js": "1.3.7",
-        "classnames": "2.2.5",
-        "codemirror": "5.18.2",
-        "colorbrewer": "1.0.0",
-        "connected-react-router": "6.3.2",
-        "create-react-class": "15.6.3",
-        "css-tree": "1.0.0-alpha24",
-        "draft-js": "0.11.0",
-        "draft-js-inline-toolbar-plugin": "3.0.0",
-        "draft-js-plugins-editor": "2.1.1",
-        "draft-js-side-toolbar-plugin": "3.0.1",
-        "draftjs-to-html": "0.8.4",
-        "element-closest": "2.0.2",
-        "es6-object-assign": "1.1.0",
-        "es6-promise": "2.3.0",
-        "eventlistener": "0.0.1",
-        "file-saver": "1.3.3",
-        "fs-extra": "3.0.1",
-        "geostyler-openlayers-parser": "1.1.4",
-        "geostyler-sld-parser": "1.3.1",
-        "history": "4.6.1",
-        "html2canvas": "0.5.0-beta4",
-        "immutable": "4.0.0-rc.12",
-        "intersection-observer": "0.7.0",
-        "intl": "1.2.2",
-        "is-equal": "1.5.5",
-        "ismobilejs": "0.5.0",
-        "istanbul-instrumenter-loader": "3.0.1",
-        "json-2-csv": "2.1.2",
-        "json-loader": "0.5.7",
-        "jszip": "3.1.5",
-        "keymirror": "0.1.1",
-        "leaflet": "1.3.1",
-        "leaflet-draw": "1.0.2",
-        "leaflet-extra-markers": "1.0.6",
-        "leaflet-minimap": "3.6.0",
-        "leaflet-plugins": "3.0.2",
-        "leaflet-simple-graticule": "1.0.2",
-        "leaflet.gridlayer.googlemutant": "0.6.4",
-        "leaflet.locatecontrol": "0.62.0",
-        "leaflet.nontiledlayer": "1.0.7",
-        "lodash": "4.17.15",
-        "lrucache": "1.0.3",
-        "moment": "2.21.0",
-        "node-geo-distance": "1.2.0",
-        "object-assign": "4.1.1",
-        "object-fit-images": "3.2.4",
-        "ogc-schemas": "2.6.1",
-        "ol": "5.3.0",
-        "pdfviewer": "0.3.2",
-        "prop-types": "15.7.2",
-        "qrcode.react": "0.9.3",
-        "react": "16.10.1",
-        "react-addons-css-transition-group": "15.6.2",
-        "react-addons-shallow-compare": "15.6.2",
-        "react-bootstrap": "0.31.0",
-        "react-checkbox-tree": "^1.5.1",
-        "react-codemirror2": "4.0.0",
-        "react-color": "2.17.0",
-        "react-confirm-button": "0.0.2",
-        "react-container-dimensions": "1.3.2",
-        "react-contenteditable": "3.3.2",
-        "react-copy-to-clipboard": "5.0.0",
-        "react-data-grid": "5.0.4",
-        "react-data-grid-addons": "5.0.4",
-        "react-dnd": "2.6.0",
-        "react-dnd-html5-backend": "2.6.0",
-        "react-dnd-test-backend": "2.6.0",
-        "react-dock": "0.2.4",
-        "react-dom": "16.10.1",
-        "react-draft-wysiwyg": "1.13.2",
-        "react-draggable": "2.2.6",
-        "react-dropzone": "3.13.1",
-        "react-error-boundary": "1.2.5",
-        "react-grid-layout": "0.16.6",
-        "react-horizontal-scrolling-menu": "^0.7.3",
-        "react-image-lightbox": "4.2.2",
-        "react-input-autosize": "1.1.4",
-        "react-intersection-observer": "8.24.1",
-        "react-intl": "2.3.0",
-        "react-notification-system": "0.2.14",
-        "react-nouislider": "2.0.1",
-        "react-overlays": "1.2.0",
-        "react-pdf": "4.0.5",
-        "react-quill": "1.1.0",
-        "react-redux": "6.0.0",
-        "react-resize-detector": "4.2.1",
-        "react-responsive": "1.3.0",
-        "react-router": "4.1.1",
-        "react-router-dom": "4.2.2",
-        "react-scroll-up": "1.3.0",
-        "react-select": "1.0.0-rc.4",
-        "react-selectize": "3.0.1",
-        "react-share": "1.15.1",
-        "react-side-effect": "1.1.0",
-        "react-sidebar": "2.3.2",
-        "react-spinkit": "2.1.2",
-        "react-swipeable-views": "0.12.2",
-        "react-textfit": "1.1.0",
-        "react-twitter-widgets": "1.3.0",
-        "react-widgets": "3.5.0",
-        "recharts": "0.22.4",
-        "recompose": "0.24.0",
-        "redux": "3.6.0",
-        "redux-logger": "3.0.6",
-        "redux-observable": "0.19.0",
-        "redux-thunk": "0.1.0",
-        "redux-undo": "0.5.0",
-        "reselect": "4.0.0",
-        "resize-observer-polyfill": "1.5.0",
-        "rxjs": "5.1.1",
-        "screenfull": "4.0.0",
-        "shpjs": "3.4.2",
-        "stickybits": "3.6.6",
-        "tinycolor2": "1.4.1",
-        "turf-bbox": "3.0.10",
-        "turf-buffer": "3.0.10",
-        "turf-intersect": "3.0.10",
-        "turf-point": "2.0.1",
-        "turf-point-on-surface": "3.0.10",
-        "turf-union": "3.0.10",
-        "url": "0.11.0",
-        "uuid": "3.0.1",
-        "vis": "4.21.0",
-        "w3c-schemas": "1.3.1",
-        "wellknown": "0.5.0",
-        "xml2js": "0.4.17"
-    },
-    "scripts": {
-        "clean": "rimraf dist",
-        "compile": "npm run clean && mkdirp ./dist && webpack --config prod-webpack.config.js --env.production",
-        "start": "webpack-dev-server --progress --colors --port 8081 --hot --inline --content-base .",
-        "postinstall": "node checkCesium.js && rimraf node_modules/react-sortable-items/node_modules/react-dom && rimraf node_modules/leaflet-simple-graticule/node_modules && rimraf node_modules/mocha && ncp node_modules/@geosolutions/mocha node_modules/mocha",
-        "test": "karma start ./karma.conf.single-run.js",
-        "continuoustest": "karma start ./karma.conf.continuous-test.js",
-        "mvntest": "karma start ./karma.conf.single-run.js --reporters junit,dots,coverage",
-        "lint": "eslint js --ext .jsx,.js",
-        "travis": "eslint js --ext .jsx,.js && karma start ./karma.conf.single-run.js --browsers FirefoxHeadless --reporters dots,junit,coverage,coveralls"
-    },
-    "author": "GeoSolutions",
-    "license": "BSD-2-Clause"
+  "name": "MapStore-C179",
+  "version": "1.0.0",
+  "description": "PUBLIACQUA",
+  "repository": "git@github.com:geosolutions-it/MapStore-C179.git",
+  "main": "index.js",
+  "eslintConfig": {
+    "extends": [
+      "@mapstore/eslint-config-mapstore"
+    ],
+    "parserOptions": {
+      "babelOptions": {
+        "configFile": "./MapStore2/build/babel.config.js"
+      }
+    }
+  },
+  "eslintIgnore": [
+    "web/client/dist/",
+    "web/client/test-resources/",
+    "web/docs/",
+    "web/client/mapstore/docs/"
+  ],
+  "browserslist": {
+    "production": [
+      ">0.5%",
+      "not dead",
+      "not op_mini all",
+      "not IE 11",
+      "not UCAndroid 12.12"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "terser": "5.18.1",
+    "@babel/core": "7.8.7",
+    "@babel/plugin-proposal-class-properties": "7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "7.8.3",
+    "@babel/preset-env": "7.8.7",
+    "@babel/preset-react": "7.8.3",
+    "@babel/runtime": "7.11.2",
+    "@geosolutions/acorn-jsx": "4.0.2",
+    "@geosolutions/jsdoc": "3.4.4",
+    "@geosolutions/mocha": "6.2.1-3",
+    "@mapstore/eslint-config-mapstore": "1.0.5",
+    "@testing-library/react": "12.1.5",
+    "axios-mock-adapter": "1.16.0",
+    "babel-loader": "8.0.5",
+    "babel-plugin-add-module-exports": "0.1.4",
+    "babel-plugin-istanbul": "6.0.0",
+    "babel-plugin-object-assign": "1.2.1",
+    "babel-plugin-react-transform": "2.0.2",
+    "babel-plugin-transform-imports": "2.0.0",
+    "copy-webpack-plugin": "5.0.2",
+    "css-loader": "5.1.2",
+    "css-minimizer-webpack-plugin": "3.0.2",
+    "denodeify": "1.2.1",
+    "docma": "1.5.1",
+    "download-cli": "1.0.1",
+    "dynamic-public-path-webpack-plugin": "1.0.4",
+    "escope": "3.2.0",
+    "eslint": "7.8.1",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-no-only-tests": "2.3.1",
+    "eslint-plugin-react": "3.3.2",
+    "expect": "1.20.1",
+    "file-loader": "2.0.0",
+    "glob": "7.1.1",
+    "html-loader": "0.5.1",
+    "html-webpack-plugin": "4.5.0",
+    "jsdoc-jsx": "0.1.0",
+    "karma": "6.4.0",
+    "karma-chrome-launcher": "3.1.1",
+    "karma-cli": "2.0.0",
+    "karma-coverage": "2.2.0",
+    "karma-coveralls": "2.1.0",
+    "karma-firefox-launcher": "2.1.2",
+    "karma-ie-launcher": "1.0.0",
+    "karma-junit-reporter": "2.0.1",
+    "karma-mocha": "2.0.1",
+    "karma-mocha-reporter": "2.2.5",
+    "karma-sourcemap-loader": "0.3.8",
+    "karma-webpack": "5.0.0",
+    "less": "4.1.1",
+    "less-loader": "8.0.0",
+    "markdownlint-cli": "0.32.2",
+    "mini-css-extract-plugin": "1.3.9",
+    "mkdirp": "0.5.1",
+    "ncp": "2.0.0",
+    "parallelshell": "1.2.0",
+    "postcss": "8.3.5",
+    "postcss-loader": "6.1.0",
+    "postcss-prefix-selector": "1.10.0",
+    "process": "0.11.10",
+    "progress-bar-webpack-plugin": "1.12.1",
+    "raw-loader": "0.5.1",
+    "react-motion": "0.5.0",
+    "react-transition-group": "1.1.3",
+    "readline-promise": "1.0.4",
+    "redux-devtools": "3.4.0",
+    "redux-devtools-dock-monitor": "1.1.2",
+    "redux-devtools-log-monitor": "1.3.0",
+    "redux-immutable-state-invariant": "2.1.0",
+    "redux-mock-store": "1.2.2",
+    "rimraf": "2.5.2",
+    "simple-git": "2.20.1",
+    "style-loader": "2.0.0",
+    "url-loader": "0.5.7",
+    "vusion-webfonts-generator": "0.4.1",
+    "webpack": "5.9.0",
+    "webpack-bundle-size-analyzer": "2.0.2",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "3.11.0"
+  },
+  "dependencies": {
+    "mapstore2": "file:MapStore2"
+  },
+  "scripts": {
+    "start": "npm run app:start",
+    "app:start": "concurrently -n frontend,backend -c green,blue \"npm:fe:start\" \"npm:be:start\"",
+    "fe:start": "webpack serve --progress --color --port 8081 --hot --inline --config webpack.config.js --content-base .",
+    "fe:start-prod": "webpack serve --progress --color --port 8081 --hot --inline --config prod.webpack.config.js --content-base .",
+    "fe:build": "npm run clean && mkdirp ./dist && webpack build --progress --color --config prod-webpack.config.js",
+    "compile": "npm run fe:build",
+    "be:build": "mvn clean install",
+    "be:start": "npm run be:build && mvn cargo:run -f ./web/pom.xml",
+    "pretest": "npm run lint",
+    "test": "karma start ./karma.conf.single-run.js --color",
+    "test:watch": "karma start ./karma.conf.continuous-test.js --color",
+    "continuoustest": "npm run test:watch",
+    "postinstall": "node MapStore2/utility/build/postInstall.js",
+    "clean": "rimraf dist",
+    "lint": "eslint js --ext .jsx,.js",
+    "updateDevDeps": "node updateDevDependencies.js"
+  },
+  "author": "GeoSolutions",
+  "license": "BSD-2-Clause"
 }
