@@ -13,8 +13,10 @@ const paths = {
 module.exports = require('./MapStore2/build/buildConfig')(
     {
         'MapStore-C179': path.join(__dirname, "js", "app"),
-        'MapStore-C179-embedded': path.join(__dirname, "MapStore2", "web", "client", "product", "embedded"),
-        'MapStore-C179-api': path.join(__dirname, "MapStore2", "web", "client", "product", "api")
+        'MapStore-C179-embedded': path.join(__dirname, "js", "embedded"),
+        'MapStore-C179-api': path.join(__dirname, "MapStore2", "web", "client", "product", "api"),
+        'geostory-embedded': path.join(__dirname, "js", "geostoryEmbedded"),
+        "dashboard-embedded": path.join(__dirname, "js", "dashboardEmbedded")
     },
     {
         "themes/default": path.join(__dirname, "themes", "theme.less")
@@ -22,17 +24,19 @@ module.exports = require('./MapStore2/build/buildConfig')(
     paths,
     extractThemesPlugin,
     true,
-    "dist/",
+    undefined,
     '.MapStore-C179',
     [
         new HtmlWebpackPlugin({
             template: path.join(__dirname, 'indexTemplate.html'),
+            publicPath: 'dist/',
             chunks: ['MapStore-C179'],
             inject: true,
             hash: true
         }),
         new HtmlWebpackPlugin({
             template: path.join(__dirname, 'embeddedTemplate.html'),
+            publicPath: 'dist/',
             chunks: ['MapStore-C179-embedded'],
             inject: true,
             hash: true,
@@ -40,13 +44,31 @@ module.exports = require('./MapStore2/build/buildConfig')(
         }),
         new HtmlWebpackPlugin({
             template: path.join(__dirname, 'apiTemplate.html'),
+            publicPath: 'dist/',
             chunks: ['MapStore-C179-api'],
-            inject: 'head',
+            inject: 'body',
             hash: true,
             filename: 'api.html'
+        }),
+        new HtmlWebpackPlugin({
+            template: path.join(__dirname, 'geostory-embedded-template.html'),
+            chunks: ['geostory-embedded'],
+            publicPath: 'dist/',
+            inject: "body",
+            hash: true,
+            filename: 'geostory-embedded.html'
+        }),
+        new HtmlWebpackPlugin({
+            template: path.join(__dirname, 'dashboard-embedded-template.html'),
+            chunks: ['dashboard-embedded'],
+            publicPath: 'dist/',
+            inject: 'body',
+            hash: true,
+            filename: 'dashboard-embedded.html'
         })
     ],
     {
+        "@mapstore/patcher": path.resolve(__dirname, "node_modules", "@mapstore", "patcher"),
         "@mapstore": path.resolve(__dirname, "MapStore2", "web", "client"),
         "@js": path.resolve(__dirname, "js")
     }
