@@ -5,8 +5,10 @@ const extractThemesPlugin = require('./MapStore2/build/themes.js').extractThemes
 module.exports = require('./MapStore2/build/buildConfig')(
     {
         'MapStore-C179': path.join(__dirname, "js", "app"),
-        'MapStore-C179-embedded': path.join(__dirname, "MapStore2", "web", "client", "product", "embedded"),
-        'MapStore-C179-api': path.join(__dirname, "MapStore2", "web", "client", "product", "api")
+        'MapStore-C179-embedded': path.join(__dirname, "js", "embedded"),
+        'MapStore-C179-api': path.join(__dirname, "MapStore2", "web", "client", "product", "api"),
+        'geostory-embedded': path.join(__dirname, "js", "geostoryEmbedded"),
+        "dashboard-embedded": path.join(__dirname, "js", "dashboardEmbedded")
     },
     {
         "themes/default": path.join(__dirname, "themes", "theme.less")
@@ -19,10 +21,11 @@ module.exports = require('./MapStore2/build/buildConfig')(
     },
     extractThemesPlugin,
     false,
-    "dist/",
+    undefined,
     '.MapStore-C179',
     [],
     {
+        "@mapstore/patcher": path.resolve(__dirname, "node_modules", "@mapstore", "patcher"),
         "@mapstore": path.resolve(__dirname, "MapStore2", "web", "client"),
         "@js": path.resolve(__dirname, "js")
     }
